requirements.txt
# Core libraries
numpy>=1.21.0
pandas>=1.3.0
matplotlib>=3.4.3
seaborn>=0.11.2
Pillow>=9.0.0
scikit-learn>=1.0.0
tqdm>=4.62.2

# PyTorch - with support for different platforms
torch>=1.12.0
torchvision>=0.13.0

# Flask web application
flask>=2.2.0
werkzeug>=2.2.0

# Optional: for dataset downloading
kaggle>=1.5.12

# For Jupyter notebooks (optional)
jupyterlab>=3.3.0
notebook>=6.4.5
ipywidgets>=7.7.0